requirements.txt
bcrypt==5.0.0
blinker==1.9.0
cffi==2.0.0
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
Flask==3.1.3
flask-cors==6.0.2
Flask-JWT-Extended==4.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
oracledb==3.4.2
pillow==12.1.1
pycparser==3.0
PyJWT==2.11.0
python-dotenv==1.2.2
typing_extensions==4.15.0
Werkzeug==3.1.6
